requirements.txt
streamlit
pydantic
PyYAML
Jinja2